--- a/chia/consensus/blockchain.py
+++ b/chia/consensus/blockchain.py
@@ -363,6 +363,7 @@ class Blockchain(BlockchainInterface):
                     )
                 else:
                     added, _ = [], []
+                await self.block_store.set_in_chain([(block_record.header_hash,)])
                 await self.block_store.set_peak(block_record.header_hash)
                 return uint32(0), uint32(0), [block_record], (added, {})
             return None, None, [], ([], {})
@@ -385,6 +386,7 @@ class Blockchain(BlockchainInterface):
 
             # Rollback sub_epoch_summaries
             self.__height_map.rollback(fork_height)
+            await self.block_store.rollback(fork_height)
 
             # Collect all blocks from fork point to new peak
             blocks_to_add: List[Tuple[FullBlock, BlockRecord]] = []
@@ -445,6 +447,8 @@ class Blockchain(BlockchainInterface):
                             if key not in hint_coin_state:
                                 hint_coin_state[key] = {}
                             hint_coin_state[key][coin_id] = lastest_coin_state[coin_id]
+
+            await self.block_store.set_in_chain([(br.header_hash,) for br in records_to_add])
 
             # Changes the peak to be the new peak
             await self.block_store.set_peak(block_record.header_hash)