This patch introduces a new `in_main_chain` column in the `full_blocks` table, indicating whether this block is part of the current heaviest chain or not.

I'm not 100% confident I'm setting this field correctly, so the only thing relying on it right now is picking a randon non-compactified block. It does improve the performance of picking this block quite significantly.

The comparison is not from the immediately prior commit (since it introduced a regression of the get_random_not_compactified() function).

The key timing metric is the `get_random_not_compactified()` function:

test  | before | after | after / before
---|---|---|---
table size | 352.5 MB | 358.8 MB | 101.8 %
Ryzen, fast SSD | 72.7s | 19.3s | 26.6 %
RPi, SD card | 376.8s  | 82.0s | 21.8 %
Xeon, HDD | 67.7s | 17.3s | 25.6 %

## raw benchmark output

This patch:

Ryzen, fast SSD:

```
version 2
50.0260s, add_full_block
53.3180s, get_full_block
5.2007s, get_full_block_bytes
52.0012s, get_full_blocks_at
7.1181s, get_block_records_by_hash
53.3044s, get_blocks_by_hash
6.2934s, get_block_record
0.7631s, get_block_records_in_range
0.0065s, get_block_records_close_to_peak
4.3690s, get_block_record
19.3385s, get_random_not_compactified
all tests completed in 251.7388s
database size: 358.810 MB
```

RPi, SD card:

```
version 2
290.5492s, add_full_block
203.2999s, get_full_block
23.6671s, get_full_block_bytes
198.3695s, get_full_blocks_at
25.9796s, get_block_records_by_hash
203.3476s, get_blocks_by_hash
25.5381s, get_block_record
3.4463s, get_block_records_in_range
0.0322s, get_block_records_close_to_peak
17.2408s, get_block_record
82.0613s, get_random_not_compactified
all tests completed in 1073.5315s
database size: 358.810 MB
```

Xeon, HDD:

```
version 2
230.2963s, add_full_block
42.8371s, get_full_block
4.0482s, get_full_block_bytes
41.6950s, get_full_blocks_at
5.0723s, get_block_records_by_hash
43.0111s, get_blocks_by_hash
4.8511s, get_block_record
0.6462s, get_block_records_in_range
0.0055s, get_block_records_close_to_peak
3.3385s, get_block_record
17.3130s, get_random_not_compactified
all tests completed in 393.1144s
database size: 358.810 MB
```

prior commit:

Ryzen, fast SSD:

```
version 2
56.3806s, add_full_block
53.0442s, get_full_block
5.1803s, get_full_block_bytes
50.1041s, get_full_blocks_at
6.8339s, get_block_records_by_hash
52.6259s, get_blocks_by_hash
6.3245s, get_block_record
0.6537s, get_block_records_in_range
0.0056s, get_block_records_close_to_peak
4.9460s, get_block_record
72.6523s, get_random_not_compactified
all tests completed in 308.7509s
database size: 352.473 MB
```

RPi, SD Card:

```
version 2
311.0953s, add_full_block
204.2302s, get_full_block
24.0304s, get_full_block_bytes
199.0710s, get_full_blocks_at
24.2053s, get_block_records_by_hash
204.2228s, get_blocks_by_hash
24.8648s, get_block_record
2.8389s, get_block_records_in_range
0.0268s, get_block_records_close_to_peak
17.6518s, get_block_record
376.7793s, get_random_not_compactified
all tests completed in 1389.0169s
database size: 352.473 MB
```

Xeon, HDD:

```
version 2
245.2260s, add_full_block
42.9780s, get_full_block
4.0179s, get_full_block_bytes
41.8684s, get_full_blocks_at
4.8848s, get_block_records_by_hash
43.1449s, get_blocks_by_hash
4.4390s, get_block_record
0.5681s, get_block_records_in_range
0.0050s, get_block_records_close_to_peak
3.3432s, get_block_record
67.6903s, get_random_not_compactified
all tests completed in 458.1655s
database size: 352.473 MB
```